2008-05-14 00:27 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/source/compiler/complex.c     ! added casting to unsigned char to force syntax error when       some unknown tokens appear. Negative values bison recognizes       as end of input.
  * harbour/contrib/rddads/adsmgmnt.c
    ! cleaned one warning

+++ harbour/contrib/rddads/adsmgmnt.c
@@ -278,7 +278,7 @@ HB_FUNC( ADSMGGETCOMMSTATS )
 
 HB_FUNC( ADSMGRESETCOMMSTATS )
 {
-   hb_retnl( s_hMgmtHandle ? AdsMgResetCommStats( s_hMgmtHandle ) : -1 );
+   hb_retnl( s_hMgmtHandle ? ( LONG ) AdsMgResetCommStats( s_hMgmtHandle ) : -1 );
 }
 
 HB_FUNC( ADSMGGETCONFIGINFO )
